Account for CFN VPCDHCPOptionsAssociation bug
When a VPCDHCPOptionsAssociation "resource" is deleted during the cleanup
phase, the result is to remove the DHCPOptions from the cluster and leave
instances without DNS. This is a bug in CloudFormation. We can work around
it by setting the DeletionPolicy on such resources to 'Retain', which will
cause CFN to ignore them rather than delete them after the stack is updated.